app: fix: wrong ping values when average is taken
The solution chosen is to return the total ping of all the packets and
let the caller handle averaging the ping value.

Other fix could be to update on the user side by considering the
number of ips for averaging the ping across ips instead of the number
of packets.