app: fix: crash when server state is some disabled
It is possible to have the server state to be some disabled when steam
adds a new ip to the region and that region was previously
disabled. So on the next startup, the server state would be some
disabled which would reach unreachable!() code. Fix is simple and self
explanatory.